Branch b1_6 b=12211 i=green, johann
suppress every allocation attempt message noise, and shut off a compile
warning on x86_64 in obd_alloc_fail(), due to size_t being unsigned long
there.

+++ lustre/lvfs/lvfs_lib.c
@@ -36,7 +36,8 @@ int obd_alloc_fail(const void *ptr, const char *name, const char *type,
         if (ptr == NULL ||
             (ll_rand() & OBD_ALLOC_FAIL_MASK) < obd_alloc_fail_rate) {
                 CERROR("%s%salloc of %s (%u bytes) failed at %s:%d\n",
-                       ptr ? "force " :"", type, name, size, file, line);
+                       ptr ? "force " :"", type, name, (unsigned int)size, file,
+                       line);
                 CERROR("%d total bytes allocated by Lustre, %d by Portals\n",
                        atomic_read(&obd_memory), atomic_read(&libcfs_kmemory));
                 return 1;
